project/pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "pytest>=9.0.3",
]

[tool.pytest.ini_options]
pythonpath = ["."]
